Feat: ê°dto 가격을 ㅑint -> bigdecimal 변경

+++ src/main/java/com/jelly/zzirit/domain/adminItem/dto/response/AdminItemResponse.java
@@ -4,6 +4,8 @@ import com.jelly.zzirit.domain.item.entity.Item;
 import com.jelly.zzirit.domain.item.entity.stock.ItemStock;
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder // record는 getter, allagrsconstructor 자동 생성
 public record AdminItemResponse (
     Long id, // record는 다 private final
@@ -11,7 +13,7 @@ public record AdminItemResponse (
     String imageUrl,
     String type,
     String brand,
-    int price,
+    BigDecimal price,
     int initialStockQuantity,
     int soldStockQuantity
 ){
@@ -22,7 +24,7 @@ public record AdminItemResponse (
                 .imageUrl(item.getImageUrl())
                 .type(item.getTypeBrand().getType().getName()) // TODO: enum으로 관리?
                 .brand(item.getTypeBrand().getBrand().getName())
-                .price(item.getPrice().intValue()) // TODO: 추후 bigdecimal로 스펙 변경 필요
+                .price(item.getPrice()) // TODO: 추후 bigdecimal로 스펙 변경 필요
                 .initialStockQuantity(itemStock.getQuantity()) // Todo: 재고수량 둘 다 줄지? 상품-재고 분리?(재고는 캐시x?), 단방향(양방향 할거면 합치기?)
                 .soldStockQuantity(itemStock.getSoldQuantity())
                 .build();
